Fixed start script
